Some objects were moved into Celeste folder in order to get rid of them in the future
