+++ src/components/Celeste/Dashboard/AccountBanner.js
@@ -3,14 +3,14 @@ import { animated, useSpring } from 'react-spring'
 import { GU, Help, LoadingRing, useTheme } from '@1hive/1hive-ui'
 
 import AccountBannerInfo from './AccountBannerInfo'
-import CircleGraph from '../CircleGraph'
-import { useCourtConfig } from '../../providers/CourtConfig'
-import { useJurorDrafted } from '../../hooks/useJurorDrafted'
-import { useJurorFirstTimeHNYActivation } from '../../hooks/useHNY'
-import { useTotalActiveBalance } from '../../hooks/useCourtStats'
+import CircleGraph from '../../CircleGraph'
+import { useCourtConfig } from '../../../providers/CourtConfig'
+import { useJurorDrafted } from '../../../hooks/useJurorDrafted'
+import { useJurorFirstTimeHNYActivation } from '../../../hooks/useHNY'
+import { useTotalActiveBalance } from '../../../hooks/useCourtStats'
 
-import { ACCOUNT_STATUS_JUROR_ACTIVE } from '../../types/account-status-types'
-import { formatUnits, getPercentageBN, bigNum } from '../../lib/math-utils'
+import { ACCOUNT_STATUS_JUROR_ACTIVE } from '../../../types/account-status-types'
+import { formatUnits, getPercentageBN, bigNum } from '../../../lib/math-utils'
 
 import hnySpringIcon from '../../assets/IconHNYSpring.svg'
 import userIcon from '../../assets/IconUser.svg'
