+++ src/components/Celeste/Dashboard/AccountBannerInfo.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { GU, textStyle, Timer, useTheme } from '@1hive/1hive-ui'
-import { useCourtClock } from '../../providers/CourtClock'
+import { useCourtClock } from '../../../providers/CourtClock'
 
 function AccountBannerInfo({ title, titleColor = '', paragraph, showTimer }) {
   const theme = useTheme()
