+++ src/components/Celeste/Dashboard/HNYLockedDistribution.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import { GU, Link, textStyle, useTheme } from '@1hive/1hive-ui'
 
-import { useCourtConfig } from '../../providers/CourtConfig'
-import { formatUnits } from '../../lib/math-utils'
+import { useCourtConfig } from '../../../providers/CourtConfig'
+import { formatUnits } from '../../../lib/math-utils'
 
 function HNYLockedDistribution({ distribution, text }) {
   const theme = useTheme()
