+++ src/components/Celeste/Dashboard/panels/WithdrawHNY.js
@@ -1,7 +1,7 @@
 import React, { useCallback } from 'react'
 import HNYForm from './HNYForm'
-import { formatUnits } from '../../../lib/math-utils'
-import { useCourtConfig } from '../../../providers/CourtConfig'
+import { formatUnits } from '../../../../lib/math-utils'
+import { useCourtConfig } from '../../../../providers/CourtConfig'
 
 const WithdrawHNY = React.memo(function WithdrawHNY({
   onWithdrawHNY,
