+++ src/lib/web3-utils.js
@@ -21,7 +21,7 @@ export function encodeFunctionData(contract, functionName, params) {
 }
 
 export function getUseWalletProviders() {
-  const providers = [{ id: 'injected' }, { id: 'frame' }]
+  const providers = [{ id: 'injected' }]
 
   if (env('FORTMATIC_API_KEY')) {
     providers.push({
